requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
